requirements.txt
amqp==5.3.1
asgiref==3.9.0
billiard==4.2.1
captcha==0.7.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
Django==5.2.4
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-shortuuidfield==0.1.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
factory_boy==3.3.3
Faker==37.4.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==11.3.0
postmarker==1.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
ruff==0.12.3
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
